Refactors api initialization to use `Channel`
- Makes index.js leaner by removing `Channel` construction there
  but leaves code which wires up the api initially.
- Renames setup.js to initialize.js and uses `Channel` instead of
  'message' event.
