+++ lib/api/index.js
@@ -1,15 +1,12 @@
-import setup from './setup'
-import Channel from './channel'
+import initializeApi from './initialize'
 import FieldLocale from './field-locale'
 import createWindow from './window'
 import createEntry from './entry'
 import createSpace from './space'
 
-export var init = setup(createWidgetAPI)
+export var init = initializeApi(createWidgetAPI)
 
-function createWidgetAPI ({id, entry, locales, field, fieldInfo}) {
-  let channel = new Channel(id, window.parent)
-
+function createWidgetAPI (channel, {entry, locales, field, fieldInfo}) {
   return {
     locales,
     field: new FieldLocale(channel, field),
