+++ lib/api/initialize.js
@@ -1,0 +1,26 @@
+import Channel from './channel'
+import Signal from './signal'
+
+export default function initializeApi (apiCreator) {
+  const channel = new Channel(null, window.parent)
+  let apiInitCallbacks = new Signal()
+  let createdApi
+
+  const removeHandler = channel.addHandler('connect', function (params) {
+    removeHandler()
+    channel.sourceId = params.id
+
+    createdApi = apiCreator(channel, params)
+
+    apiInitCallbacks.dispatch(createdApi)
+    apiInitCallbacks = null
+  })
+
+  return function init (initCb) {
+    if (createdApi) {
+      initCb(createdApi)
+    } else {
+      apiInitCallbacks.attach(initCb)
+    }
+  }
+}
